# Conflicts:
#	Example/Podfile.lock
#	Example/Pods/Local Podspecs/ForYouAndMe.podspec.json
#	Example/Pods/Manifest.lock
#	Example/Pods/Pods.xcodeproj/project.pbxproj
#	Example/Pods/Target Support Files/ForYouAndMe/ForYouAndMe-Info.plist
#	Example/Pods/Target Support Files/ForYouAndMe/ResourceBundle-ForYouAndMe-ForYouAndMe-Info.plist
#	ForYouAndMe.podspec
#	ForYouAndMe/Classes/UI/ViewControllers/PreferencesViewController.swift

+++ ForYouAndMe/Classes/Entities/Feed.swift
@@ -231,7 +231,7 @@ struct FeedResolvedTemplate: Decodable {
             } else {
                 self.resolved = resolvedString
             }
-        } else if let resolvedInt = try? container.decodeIfPresent(Int.self, forKey: .resolved) {
+        } else if let resolvedInt = try? container.decodeIfPresent(Float.self, forKey: .resolved) {
             self.resolved = "\(resolvedInt)"
         } else {
             self.resolved = nil
